requirements.txt
Flask
easyocr
Pillow
requests
